Create separate test suite for tests that traverse all the pages. Closes #1.
The problem here was that there was a single browser instance for the suite of tests and it was constantly browsing around. There is no reason to have a single browser per suite; it should instead be per test.

This commit has some nice functions using the async library for making assertions across a number of pages easily.